// web/src/types/generated/types.gen.ts
export type CourseOutput = {
    /**
     * Uid
     *
     * UID
     */
    uid: string;
    /**
     * Title
     *
     * Course title
     */
    title: string;
    /**
     * Description
     *
     * Course description
     */
    description?: string | null;
    /**
     * Order
     *
     * Display order
     */
    order: number;
    /**
     * Is Archived
     *
     * Archive flag
     */
    is_archived: boolean;
    /**
     * Created Date
     *
     * Created date
     */
    created_date: string;
    /**
     * Updated Date
     *
     * Updated date
     */
    updated_date: string;
    /**
     * Directions
     *
     * List of course directions
     */
    directions?: Array<DirectionFull>;
};
// ...
export type DirectionFull = {
    /**
     * Uid
     *
     * UID
     */
    uid: string;
    /**
     * Title
     *
     * Direction title
     */
    title: string;
    /**
     * Description
     *
     * Direction description
     */
    description?: string | null;
    /**
     * Is Archived
     *
     * Archive flag
     */
    is_archived: boolean;
    /**
     * Created Date
     *
     * Created date
     */
    created_date: string;
    /**
     * Updated Date
     *
     * Updated date
     */
    updated_date: string;
};

/**
 * direction.DirectionInput
 */
export type DirectionInput = {
    /**
     * Title
     *
     * Direction title
     */
    title: string;
    /**
     * Description
     *
     * Direction description
     */
    description?: string | null;
    /**
     * Course Uids
     *
     * Course Uids
     */
    course_uids?: Array<string> | null;
    /**
     * Is Archived
     *
     * Archive flag
     */
    is_archived?: boolean;
};

/**
 * direction.DirectionListedOutput
 */
export type DirectionListedOutput = {
    /**
     * Items
     *
     * List of directions
     */
    items: Array<DirectionOutput>;
    /**
     * Count
     *
     * Total count
     */
    count: number;
};

/**
 * direction.DirectionOutput
 */
export type DirectionOutput = {
    /**
     * Uid
     *
     * UID
     */
    uid: string;
    /**
     * Title
     *
     * Direction title
     */
    title: string;
    /**
     * Description
     *
     * Direction description
     */
    description?: string | null;
    /**
     * Is Archived
     *
     * Archive flag
     */
    is_archived: boolean;
    /**
     * Created Date
     *
     * Created date
     */
    created_date: string;
    /**
     * Updated Date
     *
     * Updated date
     */
    updated_date: string;
    /**
     * Students Count
     *
     * Count of students on direction
     */
    students_count?: number;
    /**
     * Courses
     *
     * List of direction courses
     */
    courses?: Array<CourseOutput>;
};

/**
 * direction.DirectionUpdate
 */
export type DirectionUpdate = {
    /**
     * Title
     *
     * Direction title
     */
    title?: string | null;
    /**
     * Description
     *
     * Direction description
     */
    description?: string | null;
    /**
     * Course Uids
     *
     * Course Uids
     */
    course_uids?: Array<string> | null;
    /**
     * Is Archived
     *
     * Archive flag
     */
    is_archived?: boolean | null;
};